apps/wiki-tutor/pyproject.toml
[project]
name = "wiki-tutor"
version = "0.1.0"
description = "Wiki-aware Socratic tutor TUI for the second-brain Obsidian vault"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "textual>=0.80",
    "rich>=13.7",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "websockets>=12.0",
]

[project.scripts]
wt = "wt.cli:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wt"]
